fix(tools): return a HAR parse failure instead of throwing
`parseHar` threw on an invalid upload from inside `FileReader.onload`, which
runs outside the promise chain, so the `.catch` in `loadFromFile` never saw it
and the throw reached `window.onerror`. Every bad file a person dropped on
/tools/har-viewer became a Sentry issue (UNLIGHTHOUSE-A).

It now returns a tagged `Ok` or `Err`, reads every entry field defensively, and
skips an entry with no `request` rather than failing the whole upload.
